FEM: Python Gui Base object without display modes
